POI Controller SRP fix
